Lets handle file IO errors. Don't use HW dependent size_t in shared file records.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@40203 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/trace/TraceConfigStorage.h
@@ -104,8 +104,8 @@ private:
 		tagEnd
 	};
 
-	void putItem(ITEM tag, size_t len, const void* data);
-	bool getItemLength(ITEM& tag, size_t& len);
+	void putItem(ITEM tag, ULONG len, const void* data);
+	bool getItemLength(ITEM& tag, ULONG& len);
 
 	SH_MEM_T m_handle;
 	ShMemHeader* m_base;
